Add files via upload
The original text file with the gene expression values is too large for github and must be broken up into two parts to post. There is a direct link to the file to run in the Rmarkdown script. Not fully analyzed but found at: https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE109887